BREAK: Composer update Silverstripe to 3.5

+++ framework/forms/gridfield/GridFieldExportButton.php
@@ -24,7 +24,7 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 	 * @var boolean
 	 */
 	protected $csvHasHeader = true;
-	
+
 	/**
 	 * Fragment to write the button to
 	 */
@@ -44,14 +44,15 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 	 */
 	public function getHTMLFragments($gridField) {
 		$button = new GridField_FormAction(
-			$gridField, 
-			'export', 
+			$gridField,
+			'export',
 			_t('TableListField.CSVEXPORT', 'Export to CSV'),
-			'export', 
+			'export',
 			null
 		);
 		$button->setAttribute('data-icon', 'download-csv');
-		$button->addExtraClass('no-ajax');
+		$button->addExtraClass('no-ajax action_export');
+		$button->setForm($gridField->getForm());
 		return array(
 			$this->targetFragment => '<p class="grid-csv-button">' . $button->Field() . '</p>',
 		);
@@ -90,6 +91,25 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 			return SS_HTTPRequest::send_file($fileData, $fileName, 'text/csv');
 		}
 	}
+	
+	/**
+	 * Return the columns to export
+	 * 
+	 * @param GridField $gridField 
+	 * 
+	 * @return array
+	 */
+	protected function getExportColumnsForGridField(GridField $gridField) {
+		if($this->exportColumns) {
+			$exportColumns = $this->exportColumns;
+		} else if($dataCols = $gridField->getConfig()->getComponentByType('GridFieldDataColumns')) {
+			$exportColumns = $dataCols->getDisplayFields($gridField);
+		} else {
+			$exportColumns = singleton($gridField->getModelClass())->summaryFields();
+		}
+
+		return $exportColumns;
+	}
 
 	/**
 	 * Generate export fields for CSV.
@@ -99,12 +119,8 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 	 */
 	public function generateExportFileData($gridField) {
 		$separator = $this->csvSeparator;
-		$csvColumns = ($this->exportColumns)
-			? $this->exportColumns
-			: singleton($gridField->getModelClass())->summaryFields();
+		$csvColumns = $this->getExportColumnsForGridField($gridField);
 		$fileData = '';
-		$columnData = array();
-		$fieldItems = new ArrayList();
 
 		if($this->csvHasHeader) {
 			$headers = array();
@@ -118,10 +134,10 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 			$fileData .= "\"" . implode("\"{$separator}\"", array_values($headers)) . "\"";
 			$fileData .= "\n";
 		}
-		
+
 		//Remove GridFieldPaginator as we're going to export the entire list.
 		$gridField->getConfig()->removeComponentsByType('GridFieldPaginator');
-		
+
 		$items = $gridField->getManipulatedList();
 
 		// @todo should GridFieldComponents change behaviour based on whether others are available in the config?
@@ -132,7 +148,7 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 		}
 
 		foreach($items->limit(null) as $item) {
-			if($item->hasMethod('canView') && $item->canView()) {
+			if(!$item->hasMethod('canView') || $item->canView()) {
 				$columnData = array();
 
 				foreach($csvColumns as $columnSource => $columnHeader) {
@@ -146,16 +162,23 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 						$value = $columnHeader($relObj);
 					} else {
 						$value = $gridField->getDataFieldValue($item, $columnSource);
+
+						if($value === null) {
+							$value = $gridField->getDataFieldValue($item, $columnHeader);
+						}
 					}
 
 					$value = str_replace(array("\r", "\n"), "\n", $value);
 					$columnData[] = '"' . str_replace('"', '""', $value) . '"';
 				}
+
 				$fileData .= implode($separator, $columnData);
 				$fileData .= "\n";
 			}
 
-			$item->destroy();
+			if($item->hasMethod('destroy')) {
+				$item->destroy();
+			}
 		}
 
 		return $fileData;
@@ -175,7 +198,7 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 		$this->exportColumns = $cols;
 		return $this;
 	}
-	
+
 	/**
 	 * @return string
 	 */
